v0.0.14 - resolve document field
