+++ src/scalar_option/mod.rs
@@ -3,6 +3,7 @@ use serde::{Deserialize, Serialize};
 pub mod get_from_document;
 pub mod mysql_response_row_to_input_doc;
 pub mod postgres_response_row_to_input_doc;
+pub mod resolve_document_field;
 pub mod sqlite_response_row_to_input_doc;
 pub mod to_async_graphql_value;
 pub mod to_bson_element_type;
